+++ src/main/java/br/com/agendamento/api/model/Usuario.java
@@ -30,9 +30,5 @@ public class Usuario {
         this.senha = dados.senha();
     }
 
-    public Usuario(Long idUsuario) {
-        this.idUsuario = idUsuario;
-    }
-
 
 }
