+++ src/main/java/br/com/agendamento/api/repository/UsuarioRepository.java
@@ -7,6 +7,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
     Usuario findByEmail(String email);
-
     boolean existsByEmail(String email);
 }
